픽셀 쉐이더 투명도 연산 방식 변경 (그림판 참조)

+++ JNSEngine/JNSEngine/jnsMonsterCommonInfo.h
@@ -7,7 +7,7 @@ namespace jns
 	struct MonsterCommonInfo
 	{
 		int hp;
-		int skillCoolDown;
+		float skillCoolDown;
 		int dmg;
 		int skilldmg;
